"InWork"

+++ Eclipse_Workspace/Tower Defense/src/objects/Shot.java
@@ -2,41 +2,40 @@ package objects;
 
 public class Shot {
 
-	private Double strength; // strength of shots
-	private Double speed; // number of shots in 'time'
-	private Double velocity; // velocity of shots
+	private int strength; // strength of shots
+	private int speed; // number of shots in 'time'
+	private int velocity; // velocity of shots
 
-	public Shot(Double pStrength, Double pSpeed, Double pVelocity) {
+	public Shot(int pStrength, int pSpeed, int pVelocity) {
 		this.strength = pStrength;
 		this.speed = pSpeed;
 		this.velocity = pVelocity;
 	}
 
-	public Double getStrength() // returns value of 'strength'
+	public int getStrength() // returns value of 'strength'
 	{
 		return strength;
 	}
 
-	public void setStrength(Double strength) {
+	public void setStrength(int strength) {
 		this.strength = strength;
 	}
 
-	public Double getSpeed() // returns value of 'speed'
+	public int getSpeed() // returns value of 'speed'
 	{
 		return speed;
 	}
 
-	public void setSpeed(Double speed) {
+	public void setSpeed(int speed) {
 		this.speed = speed;
 	}
 
-	public Double getVelocity() // returns value of 'velocity'
+	public int getVelocity() // returns value of 'velocity'
 	{
 		return velocity;
 	}
 
-	public void setVelocity(Double velocity) {
+	public void setVelocity(int velocity) {
 		this.velocity = velocity;
 	}
-
 }
